==> cogsci_experiments/01e. Comparison singlebranch vs multibranch.ipynb <==
import pandas as pd
import numpy as np
from isaac.statistical_tests import is_best_model_significantly_better

directory = "GRU_singlebranch/"
question_type = "mass"
stats = pd.read_hdf(directory+question_type+"_stats.h5")

sb_mass_accuracies = []
for _, seed_df in stats.groupby("seed"):
    sb_mass_accuracies.append(seed_df["Val Accuracy"].max())

print("Mass:", np.mean(sb_mass_accuracies), np.std(sb_mass_accuracies))

question_type = "force"
stats = pd.read_hdf(directory+question_type+"_stats.h5")

sb_force_accuracies = []
for _, seed_df in stats.groupby("seed"):
    sb_force_accuracies.append(seed_df["Val Accuracy"].max())

print("Force:", np.mean(sb_force_accuracies), np.std(sb_force_accuracies))

directory = "seq_length_and_res_plots/"
stats = pd.read_hdf(directory+"stats.h5")

stats = stats.query("step_size == 1 and seq_start == 0.")

mb_mass_accuracies = []
mb_force_accuracies = []

for _, seed_df in stats.groupby("seed"):
    mb_mass_accuracies.append(seed_df["Mass Val Accuracy"].max())
    mb_force_accuracies.append(seed_df["Force Val Accuracy"].max())

print("Mass", np.mean(mb_mass_accuracies), np.std(mb_mass_accuracies))
print("Force", np.mean(mb_force_accuracies), np.std(mb_force_accuracies))

is_best_model_significantly_better([("sb_mass", sb_mass_accuracies), ("mb_mass", mb_mass_accuracies)])
is_best_model_significantly_better([("sb_force", sb_force_accuracies), ("mb_force", mb_force_accuracies)])
