==> cogsci_experiments/03a. Sequence length and resolution.ipynb <==
import torch
import torch.nn as nn
import pandas as pd

from isaac.dataset import read_dataset, prepare_dataset
from isaac.constants import POSITION_COLS, MASS_CLASS_COLS, BASIC_TRAINING_COLS, FORCE_CLASS_COLS, PUCK_SQUARE_DISTANCES, PUCK_ANGLE_FEATURES
from isaac.sanity import class_proportions
from isaac.models import MultiBranchModel, initialise_model
from isaac.training import training_loop
import matplotlib.pyplot as plt
import numpy as np

data_directory = "seq_length_and_res_plots/"
model_directory = "models/seq_length_and_res/"

from isaac.utils import get_cuda_device_if_available, create_directory
device = get_cuda_device_if_available()
print(device)

create_directory(data_directory)
create_directory(model_directory)

NORMALISE_DATA = True
BATCH_SIZE = 128
EPOCHS = 100
SEQ_END = 2700

base_resolution = 60 #fps
step_sizes = [int(base_resolution/resolution) for resolution in [20, 30, 60]]

base_length = 45 #s
seq_starts = [(base_length - length_in_s)*base_resolution for length_in_s in [45, 30, 15]]

seq_starts, step_sizes

train_trials = read_dataset("data/train_passive_trials.h5", n_trials=3500, cols=BASIC_TRAINING_COLS)
val_trials = read_dataset("data/val_passive_trials.h5", n_trials=1000, cols=BASIC_TRAINING_COLS)

# TRAINING

INPUT_DIM = len(BASIC_TRAINING_COLS)    # input dimension
HIDDEN_DIM = 25  # hidden layer dimension
N_LAYERS = 4     # number of hidden layers
OUTPUT_DIM = 3   # output dimension
DROPOUT = 0.5

network_params = (INPUT_DIM, HIDDEN_DIM, OUTPUT_DIM, DROPOUT)

loaders, scaler = prepare_dataset([train_trials, val_trials], 
                                  class_columns=[list(MASS_CLASS_COLS), list(FORCE_CLASS_COLS)], 
                                  training_columns=BASIC_TRAINING_COLS, batch_size=BATCH_SIZE, 
                                  normalise_data=NORMALISE_DATA, device=device, multiclass=True)


stats_dfs = []

for step_size in step_sizes:
    for seq_start in seq_starts:
        print("step", step_size, "seq_start", seq_start)
        for seed in [0, 1, 2]:
            df = pd.DataFrame(columns=["step_size", "seq_start", "Epoch", "Mass Loss", "Force Loss", "Mass Train Accuracy",
                                       "Mass Val Accuracy", "Force Train Accuracy", "Force Val Accuracy", "seed"])

            model, error, optimizer = initialise_model(network_params, lr=0.01, seed=seed, device=device, arch=MultiBranchModel)
            epoch_losses, epoch_accuracies, [best_mass_model, best_force_model] = training_loop(model, optimizer, 
                                                                                                error, loaders[0], loaders[1], 
                                                                                                EPOCHS, seq_end=SEQ_END,
                                                                                                step_size=step_size,
                                                                                                seq_start=seq_start,
                                                                                                multibranch=True)

            torch.save(best_mass_model.state_dict(), "models/seq_length_and_res/best_mass_model_step_%d_start_%d_seed_%d.pt" % (step_size, seq_start, seed))
            torch.save(best_force_model.state_dict(), "models/seq_length_and_res/best_force_model_step_%d_start_%d_seed_%d.pt" % (step_size, seq_start, seed))

            train_accuracies = np.array(epoch_accuracies[0])
            val_accuracies = np.array(epoch_accuracies[1]) 

            df["Epoch"] = np.arange(EPOCHS)
            df["Mass Loss"] = epoch_losses[:, 0]
            df["Force Loss"] = epoch_losses[:, 1]        
            df["Mass Train Accuracy"] = train_accuracies[:, 0]
            df["Mass Val Accuracy"] = val_accuracies[:, 0]
            df["Force Train Accuracy"] = train_accuracies[:, 1]
            df["Force Val Accuracy"] = val_accuracies[:,1]
            df["step_size"] = step_size
            df["seq_start"] = seq_start            
            df["seed"] = str(seed)
            stats_dfs.append(df)
        
stats = pd.concat(stats_dfs)
stats.to_hdf(data_directory+"stats.h5", key="stats")
