==> cogsci_experiments/08. New experiment conditions.ipynb <==
import json
import pandas as pd
from isaac.constants import BASIC_TRAINING_COLS

hdf_path = "../new_exp_data/exp7_passive.h5"
trial_i = 0

for condition_id in range(1, 5):
    filename = "../new_exp_data/physics_data%d.json" % condition_id
    fd = open(filename)
    sim_data = json.load(fd)
    
    
    for sim in sim_data:
        trial = pd.DataFrame(columns=sim.keys())
        
        for key in sim.keys():
            if key in ["tR", "tM" , "practice", "refreshes", "trial_type", "condition_world_variant", "world_id"]:
                trial[key] = [sim[key]] * 2700
            else:
                trial[key] = sim[key]
                
        trial["A"] = (trial.tM == "A")
        trial["B"] = (trial.tM == "B")
        trial["same"] = (trial.tM == "same")
        
        trial["attract"] = (trial.tR == 3)
        trial["none"] = (trial.tR == 0)
        trial["repel"] = (trial.tR == -3)
        
        trial["condition"] = condition_id
                
        if trial["practice"].unique()[0]:
            continue 
            
        trial.to_hdf(path_or_buf=hdf_path, key="trial_"+str(trial_i))
        trial_i += 1

!ls -lah ../new_exp_data/

# Load participant answers

from rpy2.robjects import r, pandas2ri
pandas2ri.activate()

rdata_path = "../new_exp_data/e7_passive.rdata"
r['load'](rdata_path)

trials = r["tw"]
trials = trials.query("practice == 0")

trials.head()

count = 0
for name, df in trials.groupby(["cond_worldvar", "world_id"]):
    count += 1
    
    print(df.resp_rel.mode())
    for response in ["'attract'", "'repel'", "'none'"]:
        print(response, df.query("resp_rel == "+response).shape[0])
    
    print(df.resp_mass.mode())
    print("->", df.true_mass.unique())
    for response in ["'A'", "'B'", "'same'"]:
        print(response, df.query("resp_mass == "+response).shape[0])
    
    print()
